Chore. Code cleanup and refactor
