+++ src/BundaHubManager/Services/BundaManager.cs
@@ -1,9 +1,9 @@
 ﻿using Domain;
-using System.Collections.Generic;
+using BundaHubManager.Services.Interfaces;
 
 namespace BundaHubManager.Services
 {
-    public class BundaManager
+    public class BundaManager: IManager
     {
         private Item[] _inventory = new Item[] { }; 
         private List<Reservation> _reservations = new List<Reservation>();
@@ -22,6 +22,7 @@ namespace BundaHubManager.Services
 
             _SortInventory("name", true);
         }
+        
         private void _SortInventory(string sortBy, bool ascending = true)
         {
             switch (sortBy)
@@ -49,241 +50,40 @@ namespace BundaHubManager.Services
             }
         }
 
-        public void ViewInventory()
+        public Item[] GetInventory()
         {
-            
-            int[] colSizes = {
-                _inventory.Max(i => i.Name.Length),
-                _inventory.Max(i => i.Price.ToString().Length),
-                _inventory.Max(i => i.Quantity.ToString().Length),
-                _inventory.Max(i => i.TotalPrice.ToString().Length),
-                _inventory.Max(i => i.IsFragile.ToString().Length),
-                _inventory.Max(i => i.IsColdStored.ToString().Length),
-            };
+            // TODO: Account for reserved quantities
 
-            
-            if (colSizes[0] < "Inventory:".Length) colSizes[0] = "Inventory:".Length;
-            if (colSizes[1] < "Price".Length) colSizes[1] = "Price".Length;
-            if (colSizes[2] < "Quantity".Length) colSizes[2] = "Quantity".Length;
-            if (colSizes[3] < "Total Price".Length) colSizes[3] = "Total Price".Length;
-            if (colSizes[4] < "Fragile".Length) colSizes[4] = "Fragile".Length;
-            if (colSizes[5] < "Cold Stored".Length) colSizes[5] = "Cold Stored".Length;
-
-           
-            Console.Write("Inventory:".PadLeft(colSizes[0]));
-            Console.Write("Price".PadLeft(colSizes[1] + 2));
-            Console.Write("Quantity".PadLeft(colSizes[2] + 2));
-            Console.Write("Total Price".PadLeft(colSizes[3] + 2));
-            Console.Write("Fragile".PadLeft(colSizes[4] + 2));
-            Console.WriteLine("Cold Stored".PadLeft(colSizes[5] + 2));
-
-            
-            Console.WriteLine(new string('-', colSizes.Sum() + 10));
-
-            
-            foreach (Item item in _inventory)
-            {
-                Console.Write(item.Name.PadLeft(colSizes[0] + 2));
-                Console.Write(item.Price.ToString().PadLeft(colSizes[1] + 2));
-                Console.Write(item.Quantity.ToString().PadLeft(colSizes[2] + 2));
-                Console.Write(item.TotalPrice.ToString().PadLeft(colSizes[3] + 2));
-                Console.Write(item.IsFragile ? "Yes".PadLeft(colSizes[4] + 2) : "No".PadLeft(colSizes[4] + 2));
-                Console.WriteLine(item.IsColdStored ? "Yes".PadLeft(colSizes[5] + 2) : "No".PadLeft(colSizes[5] + 2));
-            }
-
-            
-            Console.WriteLine(new string('-', colSizes.Sum() + 10));
+            return _inventory;
         }
 
-        public void AddItem()
+        public (bool, string) AddItem(Item newItem)
         {
-            Console.WriteLine("Adding item...");
-            Console.Write("Enter item name: ");
-            string name = Console.ReadLine();
+            // TODO: Check if item already exists
 
-            Console.Write("Enter item price: ");
-            if (!decimal.TryParse(Console.ReadLine(), out decimal price))
-            {
-                Console.WriteLine("Invalid price. Item not added.");
-                return;
-            }
-
-            Console.Write("Enter item quantity: ");
-            if (!int.TryParse(Console.ReadLine(), out int quantity))
-            {
-                Console.WriteLine("Invalid quantity. Item not added.");
-                return;
-            }
-
-            bool isFragile;
-            while (true)
-            {
-                Console.Write("Is the item fragile? (yes/no): ");
-                string fragileInput = Console.ReadLine()?.Trim().ToLower();
-                if (fragileInput == "yes")
-                {
-                    isFragile = true;
-                    break;
-                }
-                else if (fragileInput == "no")
-                {
-                    isFragile = false;
-                    break;
-                }
-                Console.WriteLine("Invalid input. Please enter 'yes' or 'no'.");
-            }
-
-            bool isColdStored;
-            while (true)
-            {
-                Console.Write("Should the item be cold stored? (yes/no): ");
-                string coldStoredInput = Console.ReadLine()?.Trim().ToLower();
-                if (coldStoredInput == "yes")
-                {
-                    isColdStored = true;
-                    break;
-                }
-                else if (coldStoredInput == "no")
-                {
-                    isColdStored = false;
-                    break;
-                }
-                Console.WriteLine("Invalid input. Please enter 'yes' or 'no'.");
-            }
-
-            Item newItem = new Item(name, price, quantity, isFragile, isColdStored);
             Array.Resize(ref _inventory, _inventory.Length + 1);
             _inventory[^1] = newItem;
 
-            Console.WriteLine("Item added successfully.");
             _SortInventory("name", true);
+
+            return (true, "Item added successfully.");
         }
 
-        public void Search()
+        public (List<Reservation>, Dictionary<string, int>) GetReservations()
         {
-            Console.WriteLine("Search by (name/price/quantity): ");
-            string searchBy = Console.ReadLine()?.ToLower();
-
-            switch (searchBy)
-            {
-                case "name":
-                    Console.Write("Enter item name to search: ");
-                    string name = Console.ReadLine();
-                    var nameResults = _inventory.Where(i => i.Name.Contains(name, StringComparison.OrdinalIgnoreCase)).ToList();
-                    DisplaySearchResults(nameResults);
-                    break;
-
-                case "price":
-                    Console.Write("Enter minimum price: ");
-                    if (!decimal.TryParse(Console.ReadLine(), out decimal minPrice))
-                    {
-                        Console.WriteLine("Invalid price.");
-                        return;
-                    }
-                    Console.Write("Enter maximum price: ");
-                    if (!decimal.TryParse(Console.ReadLine(), out decimal maxPrice))
-                    {
-                        Console.WriteLine("Invalid price.");
-                        return;
-                    }
-                    var priceResults = _inventory.Where(i => i.Price >= minPrice && i.Price <= maxPrice).ToList();
-                    DisplaySearchResults(priceResults);
-                    break;
-
-                case "quantity":
-                    Console.Write("Enter minimum quantity: ");
-                    if (!int.TryParse(Console.ReadLine(), out int minQuantity))
-                    {
-                        Console.WriteLine("Invalid quantity.");
-                        return;
-                    }
-                    Console.Write("Enter maximum quantity: ");
-                    if (!int.TryParse(Console.ReadLine(), out int maxQuantity))
-                    {
-                        Console.WriteLine("Invalid quantity.");
-                        return;
-                    }
-                    var quantityResults = _inventory.Where(i => i.Quantity >= minQuantity && i.Quantity <= maxQuantity).ToList();
-                    DisplaySearchResults(quantityResults);
-                    break;
-
-
-                default:
-                    Console.WriteLine("Invalid search criteria. Choose either 'name', 'price', or 'quantity'.");
-                    break;
-
-
-            }
-
+            return (_reservations, _reservedQuantities);
         }
-        private void DisplaySearchResults(List<Item> results)
+
+        public (bool, string) AddReservation(Reservation newReservation)
         {
-            if (results.Count > 0)
-            {
-                Console.WriteLine("Search Results:");
-                foreach (var item in results)
-                {
-                    Console.WriteLine($"{item.Name}, Price: {item.Price}, Quantity: {item.Quantity}, Total Price: {item.TotalPrice}");
-                }
-            }
-            else
-            {
-                Console.WriteLine("No items found.");
-            }
-            Console.WriteLine(" ");
+            // TODO: Check reservation validity
+            // - Check if item exists
+            // - Check if quantity is available
+            // - Check if the reservation is valid
+        
+            _reservations.Add(newReservation);
+            return (true, "Reservation added successfully.");
         }
 
-        public void MakeReservation()
-        {
-            Console.Write("Enter item name to reserve: ");
-            string itemName = Console.ReadLine();
-            Console.Write("Enter quantity to reserve: ");
-            if (!int.TryParse(Console.ReadLine(), out int quantity))
-            {
-                Console.WriteLine("Invalid quantity. Reservation not made.");
-                return;
-            }
-
-            var item = _inventory.FirstOrDefault(i => i.Name.Equals(itemName, StringComparison.OrdinalIgnoreCase));
-            if (item == null)
-            {
-                Console.WriteLine("Item not found.");
-                return;
-            }
-
-            // Calculate total reserved quantity for the item
-            int totalReservedQuantity = _reservedQuantities.ContainsKey(itemName) ? _reservedQuantities[itemName] : 0;
-
-            // Calculate available quantity for reservation
-            int availableForReservation = (int)item.Quantity - totalReservedQuantity;
-
-            if (quantity > availableForReservation)
-            {
-                Console.WriteLine($"Only {availableForReservation} items available for reservation.");
-                return;
-            }
-
-            // Update reserved quantities
-            if (_reservedQuantities.ContainsKey(itemName))
-            {
-                _reservedQuantities[itemName] += quantity;
-            }
-            else
-            {
-                _reservedQuantities[itemName] = quantity;
-            }
-
-            _reservations.Add(new Reservation(itemName, quantity));
-            Console.WriteLine("Reservation made successfully.");
-        }
-
-        public void ViewReservations()
-        {
-            Console.WriteLine("Current Reservations:");
-            foreach (var reservation in _reservations)
-            {
-                Console.WriteLine($"Item: {reservation.ItemName}, Quantity: {reservation.Quantity}, Date: {reservation.ReservationDate}");
-            }
-        }
     }
 }
