+++ src/BundaHubManager/UI/Interfaces/IInterface.cs
@@ -1,0 +1,7 @@
+namespace BundaHubManager.UI.Interfaces
+{
+    public interface IInterface
+    {
+        void Run();
+    }
+}
